requirements.txt
azure-functions
azure-search-documents 
azure-storage-blob     
azure-core
requests
sendgrid

